#42871 Fixed a problem with reloading day off rules once deleted, also fixed some inconsistensies and blinking behavior

+++ Domain/Optimization/DayOffRulesModelPersister.cs
@@ -35,7 +35,8 @@ namespace Teleopti.Ccc.Domain.Optimization
 		public void Delete(Guid id)
 		{
 			var dayOffRule = _dayOffRulesRepository.Get(id);
-			_dayOffRulesRepository.Remove(dayOffRule);
+			if (dayOffRule != null)
+				_dayOffRulesRepository.Remove(dayOffRule);
 		}
 
 		private void setProperies(DayOffRules dayOffRules, DayOffRulesModel dayOffRulesModel)
